Use factories to run benches on sample graphs
