+++ benches/unionfind.rs
@@ -3,157 +3,17 @@
 extern crate test;
 extern crate petgraph;
 
-use petgraph::prelude::*;
-use petgraph::{
-    EdgeType,
-};
-use petgraph::graph::{
-    node_index,
-};
+use test::Bencher;
+
+mod factories;
+use factories::*;
 
 use petgraph::algo::{connected_components, is_cyclic_undirected, min_spanning_tree};
 
-/// Petersen A and B are isomorphic
-///
-/// http://www.dharwadker.org/tevet/isomorphism/
-const PETERSEN_A: &'static str = "
- 0 1 0 0 1 0 1 0 0 0 
- 1 0 1 0 0 0 0 1 0 0 
- 0 1 0 1 0 0 0 0 1 0 
- 0 0 1 0 1 0 0 0 0 1 
- 1 0 0 1 0 1 0 0 0 0 
- 0 0 0 0 1 0 0 1 1 0 
- 1 0 0 0 0 0 0 0 1 1 
- 0 1 0 0 0 1 0 0 0 1 
- 0 0 1 0 0 1 1 0 0 0 
- 0 0 0 1 0 0 1 1 0 0
-";
-
-const PETERSEN_B: &'static str = "
- 0 0 0 1 0 1 0 0 0 1 
- 0 0 0 1 1 0 1 0 0 0 
- 0 0 0 0 0 0 1 1 0 1 
- 1 1 0 0 0 0 0 1 0 0
- 0 1 0 0 0 0 0 0 1 1 
- 1 0 0 0 0 0 1 0 1 0 
- 0 1 1 0 0 1 0 0 0 0 
- 0 0 1 1 0 0 0 0 1 0 
- 0 0 0 0 1 1 0 1 0 0 
- 1 0 1 0 1 0 0 0 0 0
-";
-
-/// An almost full set, isomorphic
-const FULL_A: &'static str = "
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 0 1 1 1 0 1 
- 1 1 1 1 1 1 1 1 1 1
-";
-
-const FULL_B: &'static str = "
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 0 1 1 1 0 1 1 1 
- 1 1 1 1 1 1 1 1 1 1
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1 
- 1 1 1 1 1 1 1 1 1 1
-";
-
-/// Praust A and B are not isomorphic
-const PRAUST_A: &'static str = "
- 0 1 1 1 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 
- 1 0 1 1 0 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 
- 1 1 0 1 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 0 
- 1 1 1 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 
- 1 0 0 0 0 1 1 1 0 0 0 0 1 0 0 0 0 0 0 0 
- 0 1 0 0 1 0 1 1 0 0 0 0 0 1 0 0 0 0 0 0 
- 0 0 1 0 1 1 0 1 0 0 0 0 0 0 1 0 0 0 0 0 
- 0 0 0 1 1 1 1 0 0 0 0 0 0 0 0 1 0 0 0 0 
- 1 0 0 0 0 0 0 0 0 1 1 1 0 0 0 0 1 0 0 0 
- 0 1 0 0 0 0 0 0 1 0 1 1 0 0 0 0 0 1 0 0 
- 0 0 1 0 0 0 0 0 1 1 0 1 0 0 0 0 0 0 1 0 
- 0 0 0 1 0 0 0 0 1 1 1 0 0 0 0 0 0 0 0 1 
- 0 0 0 0 1 0 0 0 0 0 0 0 0 1 1 1 0 1 0 0 
- 0 0 0 0 0 1 0 0 0 0 0 0 1 0 1 1 1 0 0 0 
- 0 0 0 0 0 0 1 0 0 0 0 0 1 1 0 1 0 0 0 1 
- 0 0 0 0 0 0 0 1 0 0 0 0 1 1 1 0 0 0 1 0 
- 0 0 0 0 0 0 0 0 1 0 0 0 0 1 0 0 0 1 1 1 
- 0 0 0 0 0 0 0 0 0 1 0 0 1 0 0 0 1 0 1 1 
- 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 1 1 1 0 1 
- 0 0 0 0 0 0 0 0 0 0 0 1 0 0 1 0 1 1 1 0
-";
-
-const PRAUST_B: &'static str = "
- 0 1 1 1 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 
- 1 0 1 1 0 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 
- 1 1 0 1 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 0 
- 1 1 1 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 
- 1 0 0 0 0 1 1 1 0 0 0 0 1 0 0 0 0 0 0 0 
- 0 1 0 0 1 0 1 1 0 0 0 0 0 0 0 0 0 0 0 1 
- 0 0 1 0 1 1 0 1 0 0 0 0 0 0 1 0 0 0 0 0 
- 0 0 0 1 1 1 1 0 0 0 0 0 0 0 0 0 0 1 0 0 
- 1 0 0 0 0 0 0 0 0 1 1 1 0 0 0 0 1 0 0 0
- 0 1 0 0 0 0 0 0 1 0 1 1 0 1 0 0 0 0 0 0 
- 0 0 1 0 0 0 0 0 1 1 0 1 0 0 0 0 0 0 1 0 
- 0 0 0 1 0 0 0 0 1 1 1 0 0 0 0 1 0 0 0 0 
- 0 0 0 0 1 0 0 0 0 0 0 0 0 1 1 0 0 1 0 1 
- 0 0 0 0 0 0 0 0 0 1 0 0 1 0 0 1 1 0 1 0 
- 0 0 0 0 0 0 1 0 0 0 0 0 1 0 0 1 0 1 0 1 
- 0 0 0 0 0 0 0 0 0 0 0 1 0 1 1 0 1 0 1 0 
- 0 0 0 0 0 0 0 0 1 0 0 0 0 1 0 1 0 1 1 0 
- 0 0 0 0 0 0 0 1 0 0 0 0 1 0 1 0 1 0 0 1 
- 0 0 0 0 0 0 0 0 0 0 1 0 0 1 0 1 1 0 0 1 
- 0 0 0 0 0 1 0 0 0 0 0 0 1 0 1 0 0 1 1 0 
-";
-
-/// Parse a text adjacency matrix format into a graph
-fn parse_graph<Ty: EdgeType>(s: &str) -> Graph<(), (), Ty> {
-    let mut gr = Graph::with_capacity(0, 0);
-    let s = s.trim();
-    let lines = s.lines().filter(|l| !l.is_empty());
-    for (row, line) in lines.enumerate() {
-        for (col, word) in line.split(' ')
-                                .filter(|s| s.len() > 0)
-                                .enumerate()
-        {
-            let has_edge = word.parse::<i32>().unwrap();
-            assert!(has_edge == 0 || has_edge == 1);
-            if has_edge == 0 {
-                continue;
-            }
-            while col >= gr.node_count() || row >= gr.node_count() {
-                gr.add_node(());
-            }
-            gr.update_edge(node_index(row), node_index(col), ());
-        }
-    }
-    gr
-}
-
-/// Parse a text adjacency matrix format into a *undirected* graph
-fn str_to_ungraph(s: &str) -> Graph<(), (), Undirected> {
-    parse_graph(s)
-}
-
-/// Parse a text adjacency matrix format into a *directed* graph
-fn str_to_digraph(s: &str) -> Graph<(), (), Directed> {
-    parse_graph(s)
-}
-
 #[bench]
-fn connected_components_praust_undir_bench(bench: &mut test::Bencher) {
-    let a = str_to_ungraph(PRAUST_A);
-    let b = str_to_ungraph(PRAUST_B);
+fn connected_components_praust_undir_bench(bench: &mut Bencher) {
+    let a = ungraph().praust_a();
+    let b = ungraph().praust_b();
 
     bench.iter(|| {
         (connected_components(&a), connected_components(&b))
@@ -161,9 +21,9 @@ fn connected_components_praust_undir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn connected_components_praust_dir_bench(bench: &mut test::Bencher) {
-    let a = str_to_digraph(PRAUST_A);
-    let b = str_to_digraph(PRAUST_B);
+fn connected_components_praust_dir_bench(bench: &mut Bencher) {
+    let a = digraph().praust_a();
+    let b = digraph().praust_b();
 
     bench.iter(|| {
         (connected_components(&a), connected_components(&b))
@@ -171,9 +31,9 @@ fn connected_components_praust_dir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn connected_components_full_undir_bench(bench: &mut test::Bencher) {
-    let a = str_to_ungraph(FULL_A);
-    let b = str_to_ungraph(FULL_B);
+fn connected_components_full_undir_bench(bench: &mut Bencher) {
+    let a = ungraph().full_a();
+    let b = ungraph().full_b();
 
     bench.iter(|| {
         (connected_components(&a), connected_components(&b))
@@ -181,9 +41,9 @@ fn connected_components_full_undir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn connected_components_full_dir_bench(bench: &mut test::Bencher) {
-    let a = str_to_digraph(FULL_A);
-    let b = str_to_digraph(FULL_B);
+fn connected_components_full_dir_bench(bench: &mut Bencher) {
+    let a = digraph().full_a();
+    let b = digraph().full_b();
 
     bench.iter(|| {
         (connected_components(&a), connected_components(&b))
@@ -191,9 +51,9 @@ fn connected_components_full_dir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn connected_components_petersen_undir_bench(bench: &mut test::Bencher) {
-    let a = str_to_ungraph(PETERSEN_A);
-    let b = str_to_ungraph(PETERSEN_B);
+fn connected_components_petersen_undir_bench(bench: &mut Bencher) {
+    let a = ungraph().petersen_a();
+    let b = ungraph().petersen_b();
 
     bench.iter(|| {
         (connected_components(&a), connected_components(&b))
@@ -201,9 +61,9 @@ fn connected_components_petersen_undir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn connected_components_petersen_dir_bench(bench: &mut test::Bencher) {
-    let a = str_to_digraph(PETERSEN_A);
-    let b = str_to_digraph(PETERSEN_B);
+fn connected_components_petersen_dir_bench(bench: &mut Bencher) {
+    let a = digraph().petersen_a();
+    let b = digraph().petersen_b();
 
     bench.iter(|| {
         (connected_components(&a), connected_components(&b))
@@ -211,9 +71,9 @@ fn connected_components_petersen_dir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn is_cyclic_undirected_praust_undir_bench(bench: &mut test::Bencher) {
-    let a = str_to_ungraph(PRAUST_A);
-    let b = str_to_ungraph(PRAUST_B);
+fn is_cyclic_undirected_praust_undir_bench(bench: &mut Bencher) {
+    let a = ungraph().praust_a();
+    let b = ungraph().praust_b();
 
     bench.iter(|| {
         (is_cyclic_undirected(&a), is_cyclic_undirected(&b))
@@ -221,9 +81,9 @@ fn is_cyclic_undirected_praust_undir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn is_cyclic_undirected_praust_dir_bench(bench: &mut test::Bencher) {
-    let a = str_to_digraph(PRAUST_A);
-    let b = str_to_digraph(PRAUST_B);
+fn is_cyclic_undirected_praust_dir_bench(bench: &mut Bencher) {
+    let a = digraph().praust_a();
+    let b = digraph().praust_b();
 
     bench.iter(|| {
         (is_cyclic_undirected(&a), is_cyclic_undirected(&b))
@@ -231,9 +91,9 @@ fn is_cyclic_undirected_praust_dir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn is_cyclic_undirected_full_undir_bench(bench: &mut test::Bencher) {
-    let a = str_to_ungraph(FULL_A);
-    let b = str_to_ungraph(FULL_B);
+fn is_cyclic_undirected_full_undir_bench(bench: &mut Bencher) {
+    let a = ungraph().full_a();
+    let b = ungraph().full_b();
 
     bench.iter(|| {
         (is_cyclic_undirected(&a), is_cyclic_undirected(&b))
@@ -241,9 +101,9 @@ fn is_cyclic_undirected_full_undir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn is_cyclic_undirected_full_dir_bench(bench: &mut test::Bencher) {
-    let a = str_to_digraph(FULL_A);
-    let b = str_to_digraph(FULL_B);
+fn is_cyclic_undirected_full_dir_bench(bench: &mut Bencher) {
+    let a = digraph().full_a();
+    let b = digraph().full_b();
 
     bench.iter(|| {
         (is_cyclic_undirected(&a), is_cyclic_undirected(&b))
@@ -251,9 +111,9 @@ fn is_cyclic_undirected_full_dir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn is_cyclic_undirected_petersen_undir_bench(bench: &mut test::Bencher) {
-    let a = str_to_ungraph(PETERSEN_A);
-    let b = str_to_ungraph(PETERSEN_B);
+fn is_cyclic_undirected_petersen_undir_bench(bench: &mut Bencher) {
+    let a = ungraph().petersen_a();
+    let b = ungraph().petersen_b();
 
     bench.iter(|| {
         (is_cyclic_undirected(&a), is_cyclic_undirected(&b))
@@ -261,9 +121,9 @@ fn is_cyclic_undirected_petersen_undir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn is_cyclic_undirected_petersen_dir_bench(bench: &mut test::Bencher) {
-    let a = str_to_digraph(PETERSEN_A);
-    let b = str_to_digraph(PETERSEN_B);
+fn is_cyclic_undirected_petersen_dir_bench(bench: &mut Bencher) {
+    let a = digraph().petersen_a();
+    let b = digraph().petersen_b();
 
     bench.iter(|| {
         (is_cyclic_undirected(&a), is_cyclic_undirected(&b))
@@ -271,9 +131,9 @@ fn is_cyclic_undirected_petersen_dir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn min_spanning_tree_praust_undir_bench(bench: &mut test::Bencher) {
-    let a = str_to_ungraph(PRAUST_A);
-    let b = str_to_ungraph(PRAUST_B);
+fn min_spanning_tree_praust_undir_bench(bench: &mut Bencher) {
+    let a = ungraph().praust_a();
+    let b = ungraph().praust_b();
 
     bench.iter(|| {
         (min_spanning_tree(&a), min_spanning_tree(&b))
@@ -281,9 +141,9 @@ fn min_spanning_tree_praust_undir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn min_spanning_tree_praust_dir_bench(bench: &mut test::Bencher) {
-    let a = str_to_digraph(PRAUST_A);
-    let b = str_to_digraph(PRAUST_B);
+fn min_spanning_tree_praust_dir_bench(bench: &mut Bencher) {
+    let a = digraph().praust_a();
+    let b = digraph().praust_b();
 
     bench.iter(|| {
         (min_spanning_tree(&a), min_spanning_tree(&b))
@@ -291,9 +151,9 @@ fn min_spanning_tree_praust_dir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn min_spanning_tree_full_undir_bench(bench: &mut test::Bencher) {
-    let a = str_to_ungraph(FULL_A);
-    let b = str_to_ungraph(FULL_B);
+fn min_spanning_tree_full_undir_bench(bench: &mut Bencher) {
+    let a = ungraph().full_a();
+    let b = ungraph().full_b();
 
     bench.iter(|| {
         (min_spanning_tree(&a), min_spanning_tree(&b))
@@ -301,9 +161,9 @@ fn min_spanning_tree_full_undir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn min_spanning_tree_full_dir_bench(bench: &mut test::Bencher) {
-    let a = str_to_digraph(FULL_A);
-    let b = str_to_digraph(FULL_B);
+fn min_spanning_tree_full_dir_bench(bench: &mut Bencher) {
+    let a = digraph().full_a();
+    let b = digraph().full_b();
 
     bench.iter(|| {
         (min_spanning_tree(&a), min_spanning_tree(&b))
@@ -311,9 +171,9 @@ fn min_spanning_tree_full_dir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn min_spanning_tree_petersen_undir_bench(bench: &mut test::Bencher) {
-    let a = str_to_ungraph(PETERSEN_A);
-    let b = str_to_ungraph(PETERSEN_B);
+fn min_spanning_tree_petersen_undir_bench(bench: &mut Bencher) {
+    let a = ungraph().petersen_a();
+    let b = ungraph().petersen_b();
 
     bench.iter(|| {
         (min_spanning_tree(&a), min_spanning_tree(&b))
@@ -321,9 +181,9 @@ fn min_spanning_tree_petersen_undir_bench(bench: &mut test::Bencher) {
 }
 
 #[bench]
-fn min_spanning_tree_petersen_dir_bench(bench: &mut test::Bencher) {
-    let a = str_to_digraph(PETERSEN_A);
-    let b = str_to_digraph(PETERSEN_B);
+fn min_spanning_tree_petersen_dir_bench(bench: &mut Bencher) {
+    let a = digraph().petersen_a();
+    let b = digraph().petersen_b();
 
     bench.iter(|| {
         (min_spanning_tree(&a), min_spanning_tree(&b))
